Change input to witnesses
We need to do this to avoid exposing users moves before they're
revealed, however there are many issues with this (detailed in comments).

The examples only provide static private data which does not suit our
needs as we do multiple moves per battle that are not known at the
start.