Animations + Mock AI
Add an animation phase to combat where units move to their next stage
and then move to their target and back instead of instantly teleporting.

Damage is then done during this attack with some blood particles (can
remove later) to show the damage.

Users now set where they're going to move which shows an arrow instead
of instantly teleporting there before moves are even submitted.

Mock AI when in testing mode which moves and attacks randomly. This is
to help development as rounds on-chain take forever.