Real-time damage calculation + damage numbers + dexterity bonus
Calculate the damage per-hit using contract code to fix the issue of a
character who is attacked multiple times in a round having all the
damage being shown in the hp bar in the first hit.

Factor in weight/dexterity in damage calculations.

Damage numbers are shown to give better visual feedback.

Unit stances are now taken into account in contract damage.