Tutorial via Tooltips + Hover help text + new title screen
Add in tooltip system and uses it to construct a basic tutorial to help
new players. TODO: persist which ones are seen when we figure out a
settings store.

Add hover-over text to buttons that tweens in after a while to enable
longer descriptions.

Also changes button text color to be less white, and darkens it when
hovered over.