On-chain practice (#14)
* On-chain practice

Implements on-chain practice via modifying the `PVPArenaAPI` to react to
state changes and call the p2 methods on itself. We don't check that p1
and p2 sigs differ so this is fine for now. If we need to in the future
we can just reverse the bytes between calls.

Also adds in extra checks to validate player moves on commits.

* fixes + debugging WIP

* clean-up