Join Lobby Menu (#12)
Adds a lobby menu when you click join. The list of matches is mocked out
for now. When no games are found, it directs users to practice mode.

Practice mode is also pushed in the main menu the order is changed to
Practice, Join, Create instead of Create, Join, Practice